feat(api): align config entry value types with design doc
Replace the implementation-specific Integer/Json allow-list with the
canonical .NET-named set documented in Data-Model.md and Domain-Model.md:
String, Int32, Int64, Double, Decimal, Boolean, DateTime, DateTimeOffset,
DateOnly, TimeOnly.

Int32 and Int64 are now distinct (the previous Integer accepted any long).
Double and Decimal use invariant-culture float/number parsing. Json is
dropped — it was never part of the design and is not surfaced by any
consumer. Existing tests updated from "Integer" to "Int64"; new theories
cover the numeric happy paths and reject unsupported type names and
mistyped values.

+++ tests/GroundControl.Api.Client.Tests/ConfigEntries/ConfigEntriesClientTests.cs
@@ -123,7 +123,7 @@ public sealed class ConfigEntriesClientTests : ApiHandlerTestBase
 
         var updateRequest = new GroundControl.Api.Features.ConfigEntries.Contracts.UpdateConfigEntryRequest
         {
-            ValueType = "Integer",
+            ValueType = "Int64",
             Values = [new GroundControl.Api.Features.ConfigEntries.Contracts.ScopedValueRequest { Value = "30" }],
             IsSensitive = true,
             Description = "Connection timeout"
@@ -138,7 +138,7 @@ public sealed class ConfigEntriesClientTests : ApiHandlerTestBase
         response.StatusCode.ShouldBe(HttpStatusCode.OK);
 
         var updated = await ReadRequiredJsonAsync<ConfigEntryResponse>(response, TestCancellationToken);
-        updated.ValueType.ShouldBe("Integer");
+        updated.ValueType.ShouldBe("Int64");
         updated.IsSensitive.ShouldBeTrue();
         updated.Description.ShouldBe("Connection timeout");
         updated.Version.ShouldBe(2);
